requirements.txt
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.4.26
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==20.5
sniffio==1.3.1
typing_extensions==4.13.2